Drivers: Various fixes to the gNautilus driver
- Error messages are now printed to the LogManager
- Added handling of additional error cases
- Added support for channel units
- Miscellaneous small fixes and code cleanup

Patch contributed by g.tec medical engineering GmbH


Former-commit-id: 72db8d34b7aeb94de34bb8da77de078a9c754ec7

+++ contrib/plugins/server-drivers/gtec-gnautilus/src/ovasCConfigurationgNautilusInterface.h
@@ -51,7 +51,7 @@ namespace OpenViBEAcquisitionServer
 										 OpenViBE::boolean& rBatteryLevelEnabled,
 										 OpenViBE::boolean& rValidationIndicatorEnabled,
 										 vector<OpenViBE::uint16>& rSelectedChannels,
-										 vector<OpenViBE::uint16>& rBipolarChannels,
+										 vector<OpenViBE::int32>& rBipolarChannels,
 										 vector<OpenViBE::boolean>& rCAR,
 										 vector<OpenViBE::boolean>& rNoiseReduction
 										 );
@@ -90,7 +90,7 @@ namespace OpenViBEAcquisitionServer
 		OpenViBE::boolean& m_rBatteryLevelEnabled;
 		OpenViBE::boolean& m_rValidationIndicatorEnabled;
 		vector<OpenViBE::uint16>& m_vSelectedChannels;
-		vector<OpenViBE::uint16>& m_vBipolarChannels;
+		vector<OpenViBE::int32>& m_vBipolarChannels;
 		vector<OpenViBE::boolean>& m_vCAR;
 		vector<OpenViBE::boolean>& m_vNoiseReduction;
 		vector<OpenViBE::int32> m_vComboBoxBandpassFilterIndex;
